python/requirements.txt
numpy
scikit-learn
scipy
shapely
func-timeout
install-jdk
